Camera2: fix sawtooth in camera preview
Under certain combination of preview/screen resolution, sawtooth is
seen on preview image. In setTransformMatrix(),
scaledTextureWidth/Height is unnecessarily rounded to integers,
losing precision. Removing the rounding fixes the problem.

CRs-Fixed: 705322

+++ src/com/android/camera/PhotoUI.java
@@ -318,15 +318,11 @@ public class PhotoUI implements PieListener,
             }
         } else {
             if (width > height) {
-                scaledTextureWidth = Math.max(width,
-                        (int) (height * mAspectRatio));
-                scaledTextureHeight = Math.max(height,
-                        (int)(width / mAspectRatio));
+                scaledTextureWidth = Math.max(width, height * mAspectRatio);
+                scaledTextureHeight = Math.max(height, width / mAspectRatio);
             } else {
-                scaledTextureWidth = Math.max(width,
-                        (int) (height / mAspectRatio));
-                scaledTextureHeight = Math.max(height,
-                        (int) (width * mAspectRatio));
+                scaledTextureWidth = Math.max(width, height / mAspectRatio);
+                scaledTextureHeight = Math.max(height, width * mAspectRatio);
             }
         }
